Add new files to controller folder to clean up the models index page. Also cleaned up a couple more console statements
