+++ server/models/index.js
@@ -1,5 +1,8 @@
 var request = require('request');
 var db = require('../../db');
+var Parse = require('../controllers/parsers.js');
+var Queries = require('../controllers/queries.js');
+
 var headers = {
   'access-control-allow-origin': '*',
   'access-control-allow-methods': 'GET, POST, PUT, DELETE, OPTIONS',
@@ -8,155 +11,37 @@ var headers = {
   'Content-Type': 'application/json'
 };
 
-var parseInstagramHTML = function(username, callback) {
-  console.log('username: ', username);
-  var link = 'https://www.instagram.com/' + username + '/?hl=en';
-  var parsedData = [];
-  request({ uri: link}, function(error, response, html) {
-    if(error) {
-      console.log(error)
-    }
-
-    var index = html.indexOf('window._sharedData'), index2;
-    html = html.slice(index, html.length);
-    index = html.indexOf('<');
-    index2 = html.indexOf('{')
-    html = html.slice(index2, index - 1);
-    html = JSON.parse(html)
-
-    html.entry_data.ProfilePage[0].user.media.nodes.forEach(function(post) {
-      var obj = {};
-      obj.contentType = 'picture';
-      obj.profilePic = html.entry_data.ProfilePage[0].user.profile_pic_url;
-      obj.postPic = post.display_src;
-      obj.postContent = post.caption;
-      obj.groupMemberName = html.entry_data.ProfilePage[0].user.username;
-      obj.timeStamp = post.date;
-      obj.service = 'Instagram';
-      obj.linkToPost = 'https://www.instagram.com/p/' + post.code + '/?taken-by=' + html.entry_data.ProfilePage[0].user.username
-      obj.likes = post.likes.count;
-      obj.numberComments = post.comments.count;
-      parsedData.push(obj)
-    })
-    callback(parsedData);
-  })
-};
-
-var parseTwitterAPI = function(twitterHandle, callback) {
-  if (twitterHandle.charAt(0) === '@') {
-    twitterHandle = twitterHandle.slice(1);
-  }
-  var link = 'https://api.twitter.com/1.1/statuses/user_timeline.json\?count\=3\&screen_name\=' + twitterHandle;
-  // For security, we should clean this up later
-  request({
-    method: 'GET',
-    uri: link,
-    headers: {Authorization: 'Bearer AAAAAAAAAAAAAAAAAAAAABJLxgAAAAAAaKdnMoTibNMo2hcO%2BgAc07BbXDc%3DZ39HjkTrdPf7H3EHVeH6x8XKNKJiFAxJmvqaNMhzQDyK64vJNC'}
-  }, function(err, apiResponse) {
-    if (err) {
-      console.log('could not call twitter api', err);
-    }
-    apiResponse = JSON.parse(apiResponse.body);
-
-    var parsedResponses = [];
-
-    for (var i = 0; i < apiResponse.length; i++) {
-      parsedResponses[i] = {};
-
-      parsedResponses[i].profilePic = apiResponse[i].user.profile_image_url_https;
-      parsedResponses[i].postContent = apiResponse[i].text;
-      parsedResponses[i].contentType;
-      parsedResponses[i].groupMemberName = '<user defined name for group member>';
-      parsedResponses[i].timeStamp = apiResponse[i].created_at;
-      parsedResponses[i].service = 'Twitter';
-      parsedResponses[i].linkToPost = 'https://twitter.com/' + apiResponse[i].user.screen_name + '/status/' + apiResponse[i].id_str;
-      parsedResponses[i].retweetCount = apiResponse[i].retweet_count;
-      parsedResponses[i].likes = apiResponse[i].favorite_count;
-      parsedResponses[i].numberComments;
-    }
-    callback(parsedResponses);
-  });
-};
-
-var memberIdFinder = function(memberArray, index, req, groupId, res) {
-  if (index < memberArray.length) {
-    //check if member already in DB
-    db.dbConnection.query('SELECT id FROM members WHERE (name, facebook, instagram, twitter) = (?, ?, ?, ?)', [ memberArray[index], req.body.members[memberArray[index]].facebook, req.body.members[memberArray[index]].instagram, req.body.members[memberArray[index]].twitter ], function(err, memberId) {
-      var memberId = memberId;
-      if (err) {
-        console.log('err in member query', err);
-      } 
-      //member not in DB
-      if (memberId.length === 0) {
-        //insert member into members table
-        db.dbConnection.query('INSERT INTO members SET ?', { name: memberArray[index], facebook: req.body.members[memberArray[index]].facebook, instagram: req.body.members[memberArray[index]].instagram, twitter: req.body.members[memberArray[index]].twitter }, function(err) {
-          if (err) {
-            console.log('err in members db', err);
-          } else {
-            //query for member_id
-            db.dbConnection.query('SELECT id FROM members WHERE (name, facebook, instagram, twitter) = (?, ?, ?, ?)', [ memberArray[index], req.body.members[memberArray[index]].facebook, req.body.members[memberArray[index]].instagram, req.body.members[memberArray[index]].twitter ], function(err, idForMember) {
-              if (err) {
-                console.log('error in members query', err);
-              }
-              memberId = idForMember;
-              //add member and group to join table
-               db.dbConnection.query('INSERT INTO groups_members SET ?', { group_id: groupId, member_id: memberId[0].id }, function(err, results) {
-                 if (err) {
-                   console.log('we made it this far, what happened?');
-                 }
-                 memberIdFinder(memberArray, index + 1, req, groupId, res);
-              });
-            });          
-          }
-        });
-      //member already in DB  
-      } else {
-        //add member and group to join table
-        db.dbConnection.query('INSERT INTO groups_members SET ?', { group_id: groupId, member_id: memberId[0].id }, function(err, results) {
-          if (err) {
-            console.log('we made it this far, what happened?');
-          }
-          memberIdFinder(memberArray, index + 1, req, groupId, res);
-        });
-      }
-    });
-  } else {
-    res.end();
-  }
-};
-
-
 module.exports = {
   pages: {
     get: function(req, res) {
       var responseObj = {}
 
-      var url = req.url.slice(7);
-      var username = url.slice(0, url.indexOf('/'));
-      var groupName = url.slice(username.length + 1);
+      var parsedURL = Parse.parseURL(req.url);
+      var username = parsedURL.username;
+      var groupName = parsedURL.groupName;
       responseObj.group = groupName;
       responseObj.members = [];
 
-      db.dbConnection.query("SELECT * FROM members WHERE id IN (SELECT member_id FROM groups_members WHERE group_id = (SELECT id FROM groups WHERE name = ? AND user_id = (SELECT id FROM users WHERE username = ?)));", [groupName, username], function(err, memberInformation) {
+      db.dbConnection.query("SELECT * FROM members WHERE id IN (SELECT member_id FROM groups_members WHERE group_id = (SELECT id FROM groups WHERE name = ? AND user_id = (SELECT id FROM users WHERE username = ?)));", [groupName, username], function(err, groupMemberAccountInformation) {
         if (err) {
           console.log('could not find the group member names', err);
         }
 
-        var socialMediaCompiler = function(memberInformation, index, req, res) {
-          if (index < memberInformation.length) {
+        var socialMediaCompiler = function(groupMemberAccountInformation, index, req, res) {
+          if (index < groupMemberAccountInformation.length) {
             var memberObj = {};
-            memberObj.name = memberInformation[index].name;
+            memberObj.name = groupMemberAccountInformation[index].name;
 
             // call to instagram
-            parseInstagramHTML(memberInformation[index].instagram, function(instagramData) {
+            Parse.parseInstagramHTML(groupMemberAccountInformation[index].instagram, function(instagramData) {
               memberObj.instagram = instagramData;
 
               // call to twitter
-              parseTwitterAPI(memberInformation[index].twitter, function(twitterData) {
+              Parse.parseTwitterAPI(groupMemberAccountInformation[index].twitter, function(twitterData) {
                 memberObj.twitter = twitterData;
       
                 responseObj.members.push(memberObj);
-                socialMediaCompiler(memberInformation, index + 1, req, res);
+                socialMediaCompiler(groupMemberAccountInformation, index + 1, req, res);
               });
             })
                   // call to facebook
@@ -164,12 +49,11 @@ module.exports = {
           }
 
           else {
-            console.log('responseObj', responseObj.members[0].twitter);
             res.end(JSON.stringify(responseObj));
           }
         }
 
-        socialMediaCompiler(memberInformation, 0, req, res)
+        socialMediaCompiler(groupMemberAccountInformation, 0, req, res)
       });
     },
     post: function(req, res) {
@@ -177,7 +61,6 @@ module.exports = {
       db.dbConnection.query('SELECT id FROM users WHERE username = ?', [req.body.username], function(err, userId) {
         //username in DB
         if (userId.length > 0) {
-          //make sure user does not already have this group
           //query for group_id
           db.dbConnection.query('SELECT id FROM groups WHERE (user_id, name) = (?, ?)', [ userId[0].id, req.body.groupName ], function(err, groupId) {
             var groupId = groupId;
@@ -204,7 +87,7 @@ module.exports = {
                       memberContainerArray.push(member);
                     }
 
-                    memberIdFinder(memberContainerArray, 0, req, groupId[0].id, res);
+                    Queries.memberIdFinder(memberContainerArray, 0, groupId[0].id, req, res);
                   });
                 }
               });
@@ -253,7 +136,7 @@ module.exports = {
                   memberContainerArray.push(member);
                 }
 
-                memberIdFinder(memberContainerArray, 0, req, groupId[0].id, res);
+                Queries.memberIdFinder(memberContainerArray, 0, groupId[0].id, req, res);
               });
             });
           });
@@ -324,7 +207,6 @@ module.exports = {
       });
     }
   }
-
 
 };
